+++ test/fixtures/import/circular.js
@@ -1,0 +1,5 @@
+import css from './circular.css';
+
+__export__ = css.toString();
+
+export default css;