+++ test/import-option.test.js
@@ -55,6 +55,20 @@ describe('"import" option', () => {
     expect(getErrors(stats)).toMatchSnapshot("errors");
   });
 
+  it("should work with circular `@import`", async () => {
+    const compiler = getCompiler("./import/circular.js");
+    const stats = await compile(compiler);
+
+    expect(getModuleSource("./import/circular.css", stats)).toMatchSnapshot(
+      "module"
+    );
+    expect(getExecutedCode("main.bundle.js", compiler, stats)).toMatchSnapshot(
+      "result"
+    );
+    expect(getWarnings(stats)).toMatchSnapshot("warnings");
+    expect(getErrors(stats)).toMatchSnapshot("errors");
+  });
+
   it("should work with import.filter", async () => {
     const compiler = getCompiler("./import/import.js", {
       import: {